Added an Audio Noise track generator
Generate a noise track that depends on some parameters from the related country.

In "generate.py", I added a function called by main() that generates a noise track of the same length than the voice track of a given country. The noise comes from a random noise generator, goes through a filter which cutoff frequency depends on the length of the spoken text, and a Gaussian envelope.

In "run.sh", I just copied the code that concatenates the bits of voice tracks, and do the same for the noise tracks so we get two audiofiles, one for the voices and one for the noise, they simply need to be joint together for the final stream. I let you do it.